+++ rosen/modules/render_service_client/core/ui_effect/filter/include/filter_frosted_glass_para.h
@@ -333,12 +333,12 @@ public:
         return materialColor_;
     }
 
-    void SetColorBlendMode(float colorBlendMode)
+    void SetColorBlendMode(int colorBlendMode)
     {
         colorBlendMode_ = colorBlendMode;
     }
 
-    float GetColorBlendMode() const
+    int GetColorBlendMode() const
     {
         return colorBlendMode_;
     }
@@ -466,7 +466,7 @@ private:
     bool baseVibrancyEnabled_ = true;
     float baseMaterialType_ = 0.0f;
     Vector4f materialColor_ = Vector4f(0.0f, 0.0f, 0.0f, 0.0f);
-    float colorBlendMode_ = 0.0f; // 0: LINEAR_MIX, 1: TILTED_GLASS
+    int colorBlendMode_ = 0; // 0: LINEAR_MIX, 1: TILTED_GLASS
     float vibrancyStrength_ = 0.0f;
     Vector3f lumaParams_ = Vector3f(0.0f, 0.0f, 0.0f);
     float materialColorFraction_ = 0.0f;
